Backend/requirements.txt
accelerate==0.27.2
aiohttp==3.7.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
async-timeout==3.0.1
attrs==23.2.0
auto_gptq==0.7.0.dev0+cu121
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bitsandbytes==0.43.0
blinker==1.7.0
bson==0.5.10
build==1.0.3
cachetools==5.3.2
certifi==2019.9.11
chardet==3.0.4
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
commonmark==0.9.1
datasets==2.17.0
Deprecated==1.2.14
dill==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
einops==0.7.0
fastapi==0.109.2
filelock==3.13.1
FlagEmbedding==1.2.5
Flask==3.0.2
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.10.0
gekko==1.0.6
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.27.0
google-generativeai==0.4.1
googleapis-common-protos==1.62.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.20.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
kubernetes==29.0.0
MarkupSafe==2.1.5
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mysqlclient==2.2.4
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.0
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
optimum==1.16.2
overrides==7.7.0
packaging==23.2
pandas==2.2.0
peft==0.8.2
pillow==10.2.0
posthog==3.4.1
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.8
pulsar-client==3.4.0
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.17.2
PyMuPDF==1.23.22
PyMuPDFb==1.23.22
PyPika==0.48.9
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rake-nltk==1.0.6
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==10.14.0
rouge==1.0.1
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
sentencepiece==0.2.0
setuptools==69.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
torch==2.2.0+cu121
tqdm==4.66.2
transformers==4.37.2
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
uvicorn==0.27.1
watchfiles==0.21.0
websocket-client==1.7.0
websockets==9.1
Werkzeug==3.0.1
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
